+++ CHANGELOG.md
@@ -25,6 +25,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 * fix(themes): missing tokyo-night-dark theme (https://github.com/zellij-org/zellij/pull/1972)
 * refactor(plugins): fix plugin loading data flow (https://github.com/zellij-org/zellij/pull/1995)
 * refactor(messaging): reduce extraneous cross-thread messaging (https://github.com/zellij-org/zellij/pull/1996)
+* errors: preserve caller location in `to_log` (https://github.com/zellij-org/zellij/pull/1994)
+* feat: show loading screen on startup (https://github.com/zellij-org/zellij/pull/1997)
 
 ## [0.33.0] - 2022-11-10